test(auth-commands): split test_web_login into logical modules
The web-login test module grouped 14 tests. Split it by concern into three
sibling modules: the happy path (`test_web_login`), `--scope` handling
(`test_web_login_scope`), and the error paths — the `From<WebLoginFlowError>`
arms, the `auth.ini` read error, and the control-character rejection
(`test_web_login_errors`).

Addresses review feedback on pnpm/pnpm#12851.
